+++ http4k-serverless-lambda/src/test/kotlin/org/http4k/serverless/TestApp.kt
@@ -6,8 +6,6 @@ import org.http4k.core.HttpHandler
 import org.http4k.core.RequestContexts
 import org.http4k.core.Response
 import org.http4k.core.Status.Companion.CREATED
-import org.http4k.serverless.lambda.LAMBDA_CONTEXT_KEY
-import org.http4k.serverless.lambda.LAMBDA_REQUEST_KEY
 
 object TestApp : AppLoader {
     override fun invoke(env: Map<String, String>): HttpHandler = { request ->